# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kids-r-us",
        "description": "E-commerce web api",
        "contact": {
            "name": "lleraabi@gmail.com",
            "email": "lleraabi@gmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/users/{user}": {
            "get": {
                "summary": "Obtiene un usuario por el id.",
                "description": "Obtiene un usuario por el id. Este enpoint requiere permisos de administrador",
                "operationId": "d3a14968f8b7a072620c75c7c78ed6f8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Elimina un usuario de la base de datos.",
                "description": "Elimina un usuario de la base de datos. Este endpoint requiere permisos de administrador.",
                "operationId": "e0f9a385a1d6788b99e519bd98cf7239",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id del usaurio",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/users/{user}/assign-role": {
            "patch": {
                "summary": "Asigna un rol dado a un usuario",
                "description": "Asigna un rol dado a un usuario. Este endpoint requiere permisos de administrador.",
                "operationId": "5ad0b414220edbc1b6b199835bba2bd5",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id del usaurio",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "description": "Rol a asignar",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "role": "edit"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/users/{user}/remove-role": {
            "patch": {
                "summary": "Elimina un rol a un usaurio.",
                "description": "elimina un rol a un usuario. Este endpoint requiere permisos de administrador. No se puede\n     *                  eliminar un rol al usaurio autenticado.",
                "operationId": "f870b1d58f1986675b3f3ad3d7f109ad",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id del usaurio",
                        "required": true,
                        "allowEmptyValue": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "description": "Rol a asignar",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "role": "edit"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        }
    },
...
}